search allele frequency among populations and gene annotation information

using genomic postion and ref/alt information to query gene annotation and allele frequency

pop_freq(dat_ori, pop = "db_gnomAD_exome_freq")

Arguments

dat_ori
a dataframe including chromosome, start and end postion, ref/alt nucleotide

pop
selection of table. Default is db_gnomAD_exome_freq

Value

a new dataset with allele frequency and annotation information

Examples

dat<-read.csv(system.file("extdata",
                          "anno_freq_data.csv",
                          package = "MRAT"),
              stringsAsFactors = FALSE, encoding = "UTF-8", row.names = NULL, sep = ",")
pop_dat<-pop_freq(dat, pop="db_gnomAD_exome_freq")